Add reconstruction filter weighting. Initialize left border region of histogram with 1. Update histogram/integration tests.

+++ include/mitsuba/render/histogram.h
@@ -95,10 +95,10 @@ public:
     const DynamicBuffer<Float> &data() const { return m_data; }
 
     /// Return the underlying counts for every bin
-    DynamicBuffer<UInt32> &counts() { return m_counts; }
+    DynamicBuffer<Float> &counts() { return m_counts; }
 
     /// Return the underlying counts for every bin (const version)
-    const DynamicBuffer<UInt32> &counts() const { return m_counts; }
+    const DynamicBuffer<Float> &counts() const { return m_counts; }
 
     //! @}
     // =============================================================
@@ -117,7 +117,7 @@ protected:
     ScalarPoint2i m_offset;
     int m_border_size;
     DynamicBuffer<Float> m_data;
-    DynamicBuffer<UInt32> m_counts;
+    DynamicBuffer<Float> m_counts;
     const ReconstructionFilter *m_filter;
     Float *m_weights;
 };
